+++ src/tool.h
@@ -23,7 +23,7 @@ void printRawChar(char,FILE*);
 char* doubleToString(double);
 double stringToDouble(const char*);
 char* intToString(long);
-int32_t stringToInt(const char*);
+int64_t stringToInt(const char*);
 char stringToChar(const char*);
 void errors(int);
 pair* createPair(pair*);
